Refactor Money to int and TDD to Classic Stubs
- Changed OrderDTO totalAmount to int (cents)
- Updated OrderMapper to calculate total in cents
- Created BaselinkerClientStub for better test isolation
- Refactored OrderFetchServiceTest to use Stub instead of Mocks
- Cleaned up test comments
- Added make test-contract target

+++ src/DTO/OrderDTO.php
@@ -10,7 +10,7 @@ final readonly class OrderDTO
         public string $externalId,
         public string $marketplace,
         public string $customerName,
-        public float $totalAmount,
+        public int $totalAmount,
         public \DateTimeImmutable $createdAt,
         /** @var array<int, array<string, mixed>> */
         public array $products = [],
